This fixes the "Join Now!" button so that when you land on the account creation page, you're also in the room you were in in the embedded client.
This highlights the wacky "room reset" bug that's been annoying me (and some people on the forums)
for the last few weeks.  I had thought it might be a Firefox 3 bug, but it's showing up in FF2 as
well, so it's clearly something we broke.  I'll add it to my list.