Like most truly annoying bugs, several hours of debugging distills down to basically a one line fix here.  This is a fix for the bug that causes the client to reload if you landed on a flash client URL first (e.g. #world-m5), and then went to a GWT page next (Me, Shop, etc).  This happens especially often if you move on to whirled.com from an embedded client.
Now a little background:  For reasons unknown, setting the window to not scroll 
(Window.enableScrolling(false)) will cause the flash client to reload in Firefox (didn't check IE),
if it has been loaded.  This is probably a browser-internal thing, not necessarily caused by GWT.

We were leaving the browser default active on application init (scrolling enabled), and only 
disabling it when we went to a page that does custom scrolling (which is actually true of most 
pages in our app).  In the case where you start on a GWT page and then go to a Flash URL, that's 
fine, because you don't then muck with the browser scrolling after the client is displayed.

This change merely turns off scrolling by default - pages that don't do custom scrolling get it
enabled explicitly, which was already happening.  Currently, the only page that doesn't do 
custom scrolling is the landing page.

If someone goes to the landing page, and then (without logging in or anything), follows a bookmark
to a room, then goes to a GWT page, they'll still tickle this bug.  That is such an unlikely 
set of events that I'm not going to muck with it now.  The solution would probably be to turn off 
window scrolling just before loading the flash client every time.  However, it's time to catch a 
train, this commit sets up the default that is used in almost every case anyway, and works in every 
realistic scenario.  So this is where I leave it for now.