A few bits: - generalized the member object forwarding system - implemented multi-server compatible pet following - moved support+ check into base scene access checking so that we don't duplicate that additional code in three places.
Pet following isn't totally done, if you tell your pet to go home they don't
actually go home because home might be on another server and making that work
properly is complicated. There is also an edge case where a pet is forwarded to
another server but the member never shows up that need to be dealt with.

The new behavior for pets is to go into following mode when you call them which
differs from the old mode which was to "let them out" in the room you were in
(which would not work for rooms in which you didn't have access). Now you can
call them any time and you have to take an additional step to tell them to
"stay" in the room in question (which you might not have access to do). This
seems to mostly make sense but you get an error message if you try to call two
pets in a row because you can only have one pet following you. I think that's
probably the best way, so that players learn that they have to tell a pet to
stay if they want it to stay.