Our SubItem system is all weird.  Why should a game creator have to list the trophy sources in the catalog?  The reason is that that's how we clone things, but it makes no sense from the creators standpoint.  We should either rejigger cloning (el fiasco), or auto-list sub items (maybe duable).
Anywho, lets complain if the player attempts to list a sub item whose parent is not listed, as this
puts the system into a crazy broken state from which it is quite difficult, if not impossible 
to recover from (without blowing away everything and starting over).

I believe this fixes the issue where uploading and testing trophies before listing the original
game in the catalog made the trophies not show up.  Although I'm not entirely sure I understood
that issue, so more testing is required.