Scale our payout down if a game is completed too fast. We allow a game to complete within 80% of the normal average time and still get a full payout, then we scale the payout down linearly, with a 20% bonus until we get down to 20% at which point we scale the time bonus as well so that 0 seconds is 0 payout.
This should prevent situations like we recently saw where a player discovers an
exploit that causes a game to complete in a very short time (2 seconds in the
observed case) and award them a full allotment of flow based on the game's
average time. This in turn drives a game's payout factor into the dust because
it is now seen by the system as awarding flow in a highly inequitable manner
and a fair number of non-exploiting players then have to play the game and get
crappy payouts to restore the game's payout factor to something sensible.

Now rapid fire exploits like this will pay out paltry sums of flow. It will
still, of course, be possible to discover an exploit (or simply write a game
with a built-in exploit) that will allow you to farm flow, you'll just have to
earn it in the same amount of time it would take a good player to earn it by
actually playing the game. We'll also aim to catch exploiters of this sort by
our much maligned humanity calculations and at a grosser level, by the
forthcoming per-day game flow earnings limits.