Step 1 towards the new avatar viewer: the refactor. I ended up trashing my work from last week (trying to split things 3 ways) and just split things two ways: The base room view is now RoomView, it is extended by RoomObjectView, which adds RoomObject-specific stuff as well as chatting and editing. The original division of AbstractRoomView and RoomView has long been irrelevant and so is gone. Most things should be added to RoomView unless they are specific to a networked room. I'll be creating something like RoomStudioView for an unconnected playspace in which to view avatars.
This is still messy, I didn't prune unneeded imports or clean up every little
thing. More work is coming, but "check in early, check in often" and everything
works.

Some notes:
- Don't be afraid to lazy-initialize! The room editor, scene snapshotter, and
door target controller were *always around* even if the user never used them.
- I shouldn't have to say this, but don't copy/paste code! There were some
things related to room properties that were just lifted from MsoySprite and
copied to AVRG, and a bunch of code in AVRG backends that did the same buncha
lines of code over and over instead of breaking them out into a helper method.

Not all of this has been cleaned up yet, and I will probably be going through
again and moving some things to one side or another, as well as making
the things that entities can change be a little saner: sometimes
the RoomView handles them, sometimes the RoomController, it's just
random depending on who made the change and which way the wind was blowing that
day.