The NPE in MsoySceneDirector in server switches after logging into an embedded client was actually the result of the client rapidly bouncing back and forth between servers, and eventually throwing its hands up in disgust.  I think this may have actually resolved into something unexpected (rather than throwing an error) in the higher latency connection to the public servers.
Anway, we were saving the scene we were in on logon, and doing a moveTo() on that scene once logon
was complete so that logging on got us into the same scene as we were in.  Then when moving to a 
new server, we were logging on again (to the new server), and attempting to go back to the 
previously remembered scene (on the other server).  So, we just need to clear it out once we've 
used it once.