Bring in the fancy Annotations!  Yay!
First this was just uncommenting all the lines that looked like this:

    // @Overrides // from Page

Then I (re)discovered that for GWT, we're re-implementing some classes to make the GWT compiler
happy.  The one I attack in this commit is Name - I was able to copy the Name implementation from 
narya, and it compiles, with Comparable<Name> and all.  I did have to remove an unused import to
java.regex.Pattern.

If the narya sources were still in the msoy tree, I could reference Narya directly... but since we
mavenized things, that's not possible, and GWT needs source files, not jars.  I don't like that we
have a second copy of Name sitting there that everybody will forget about, but I don't see that
we have much choice without loading narya sources as externals to msoy again.

Also, I got rid of two unchecked warnings in the server-side build in Item, as I was able to 
add type checking to the two HashMaps in there.