Returning IsSerializable from a service method was causing the GWT compiler to analyze *all* subclasses of IsSerializable, which was slowing it down, spewing that big warning a few of us have noticed, and potentially (but probably not?) increasing the GWT client download size.
Fixed by creating a little custom class for handling either of the two objects that that service
function really wanted to be able to return.  I wanted to check if this saved us any size in the
client download, but the GWT compiler output consists of a whole bunch of different junk, so I'm 
really not sure.