Fixed feeds (I think).  From the javadocs for Java 1.5, Arrays.asList() returns a list that is fixed size.  I susepect in a real Java app, calling remove() on the Iterator for such a list would throw an UnsupportedOperationException.  In GWT 1.4, it all worked fine.  In GWT 1.5 it causes the javascript execution to stop, without spewing a warning of any kind.  This makes a very simple feed I magicked up on my local machine work again.  Will need to see it in the much more robust prodcution environment to know for sure.
While I was in here testing things, I did a bunch of genericizing.  It may look like I added a 
constraint on aggregation that only supports FriendFeedMessages.  In fact, only friend feed messages
were being supported, but in a much less obvious way - via unchecked type casts.  This makes it
obvious, so that if we ever do try to aggregate something else (if we ever have something that it 
makes sense to aggregate), it'll be clear that some things need to change to support them.