One of the last big generics GWT 1.5 conversions.  This one takes care of all of the uses of AsyncCallback in services.  I find it odd that the compiler allows calling a function that requires a specific version of a generic class with a non-generic-aware instance.  In other words, the GWT 1.5 compiler lets you call
public void foo (Bar<Baz> myBar) { ... }

with 

Bar bar = new Bar();
foo(bar);

Which should be illegal, as I understand generics in Java 1.5.  This has made the conversion easier,
but it would be nice if I could at least throw a switch to have the compiler warn me about these.  
The only cases that I know this is still being done in GWT code is when a ServiceBackedDataModel 
calls a service passing itself as the callback.

There are two large bits of generic conversion left to do, and I plan to tackle them when I need 
a way to escape boring inlaws or distant relations while on vacation:
 * There are probably still Lists and Maps spread all over the place that need generics (trivial but
   time consuming)
 * DataModel and all of its implementors.  This will be relatively complicated and touches bits all 
   over our app.