Update your online friends when you change you profile info (display name, photo, status only), if you are also online (and they will therefore care).
An oddity: in MemberNodeActions.FriendEntryUpdate, I tried storing a local FriendEntry and sending
that in the NodeAction.  To make that work, I had to make MemberName and FriendEntry 
Serializable, no big deal.  However, the receiving server ended up with a FriendEntry that had a
null MemberName.  Weird.  So, now it just sends the fields it cares about and constructs a 
new FriendEntry on each server.