Managers (MemberManager, FriendManager, etc.) are fundamentally distributed object runtime entities. They deal in distributed object stuff and they come along with the implicit assumption that methods called on them are being called on the distributed object (event) thread.
I introduced @EventThread, @BlockingThread and @AnyThread a short while ago to
make that assumption a little more clear and also to allow certain methods in
managers to be marked as deviating from the norm. However, deviation is a
dangerous thing because we all just plow along with our implicit assumptions
and pay little or no attention to something as small as an annotation at the
top of a class.

Now that we have servlet code and distributed object runtime code both running
in the same server and mingling dangerously closely, we've been trying to
follow a certain pattern:

*Manager contains only distributed object code, and always uses the invoker to
do database things.

*Servlet contains only blocking code and can access the database directly and
must post runnables to the dobjmgr thread if they want to have any impact on
the distributed object world.

That's all fine and dandy, but *Servlet classes are implementing a specific RPC
interface and are not well suited as repositories for shared code. Thus if
ItemServlet and MemberServlet both need to do the same thing, there's no good
place for the code to live.

I originally created ServletUtil which then expanded into ItemUtil, ForumUtil
and so on, but *Util classes imply that they contain only static methods that
do simple things in isolation with dependencies that are passed into the method
call every time. This is not a good solution either. I also considered just
putting the shared code into the associated repository, but I don't like the
idea of sullying repositories with code not related to reading and writing from
the database.

What we really need are *Manager equivalents for blocking thread code. With
that in mind, I've been pondering what in the hell to call them for the last
few days. We're already using manager, servlet, util, repository, service,
controller, object, info, data, and half a dozen other potential candidates.
Plus many of those don't express what I want anyway.

So I've decided on *Logic since it hints at what's going on to anyone
unfortunate enough to be familiar with the term business logic, and the only
other place we're already using it is in the Game Gardens game mini-
architecture (where of course it means something very different, but oh well).

Thus if you have code that's going to be used by multiple servlet classes, put
it in a *Logic class and inject it into the servlet that needs it. Similarly if
there is code that is shared between a manager and a servlet, then you should
probably put it in a *Logic class and have the servlet call it directly and the
manager call it via the invoker thread.

I'm going to revamp all of the existing web.server.*Util classes shortly and
put them in the proper packages.