WIP towards comprehensive tracking of players and their affiliate referrals across various sessions and entry vectors.
Each visitor to our site gets a tracking number for A/B tests, and optional 
affiliate referral information, on the first visit. This gets saved across
visits: in the browser cookie (so we can track their visits to GWT pages
without being logged in), in the flash cookie (so we can track embedded games
and room visits), and on the MemberObject. Affiliate information can be 
specified in GWT page URLs, or via flash parameters on the embedded player.

If the guest leaves and then returns at a later date, their old referral info
will be loaded up from cookies, if it's available, and reported to the server. 
Once the guest converts to a regular player, their referral info will also be
stored permanently in the DB, and used as the authoritative version to clobber
their cookies whenever they log in. 

-- 

This implementation does most of the above, but is missing two features,
to be added in later checkins:
- When a guest player registers, we need to copy their referral info to the DB
- When a registered player logs off, we need to clear their referral cookies