- Mogrified MemberStatUtil into StatLogic. Please strive not to introduce new   static dependencies. Everything we need to do can now be done without them.
- Changed the way stats are updated by servlets to use the new Vilya Stat
  system's StatModifier which simplifies life on the MSOY side.

- Revamped a bunch of friend related stuff because it was wonky. MemberLogic
  now handles establishing and clearing friendships; we no longer duplicate
  that code in WebUserServlet and MemberServlet; we no longer have a NodeAction
  that turns around and loads something out of the database (for future
  reference: make the caller look things up before creating the node action);
  friendship stats are properly tracked because it happens in only one place
  (instead of happening in MemberServlet but not in WebUserServlet); removed
  dependencies from MemberRepository (more on that below).

A note on dependencies: repositories should not have any dependencies. We
previously allowed them to have a dependency on MsoyEventLog but that's a bad
idea which is no longer necessary now that we have the *Logic construct.

If you find that you need to record something to the event log in a repository,
create a logic method that calls the repository and records the event. Don't
call the log directly from the repository. *Definitely* do not make a
repository depend on a logic class like Tim did with MemberStatUtil. Tim has an
excuse however, he's never done anything like this before, so we won't require
a beating (this time).