I can't believe this has been in this state since flash r124 came out.  Quick, if you test on multiple servers locally, raise your hand. :)  I tend to frequently, and I thought I'd upgraded to flash r124 when it came out, but discovered the other day that I was still using r115.  The difference is that r124 enforces the stricter security stuff that we had to implement the PolicyServer for.  On developer machines, msoy1 starts up the PolicyServer automatically, in order to simplify our lives.  Unfortunately, that means that the server started by msoy1 was only allowing the ports that msoy1 knew about... so if you tried to connect to a local msoy2 with a r124 flash player, the security policy said "nope, not allowed" because msoy2's ports weren't being added to the policy XML.
I started down the path of adding a msoy-server.properties property to let developers specify the 
number of nodes they'd like to assume is running locally.  I abandoned that route for a couple of 
reasons: Zell pointed out that this is a silly hack and dumping stuff in the properties files isn't
really a great idea, also I'd thought it was ok because I thought it wouldn't require a production
config update, but it turns out that it did because of how DeploymentConfig.java is generated in the
build process.

Anyway, this approach is probably better.  In dev deployments, when another node connects to msoy1,
msoy1 will restart it's policy server with the ports that that node should be listening to, as 
configured in msoy-server.properties.  I wanted to just tell the MsoyPolicyServer the new ports
and have it just regenerate the XML it spits out (and in fact implemented all of that), but then
discovered that MsoyServer doesn't actually get a reference to the MsoyPolicyServer, but merely a
mina object that has the shutdown() method on it.  I thought about refactoring to have the 
PolicyServer implement a shutdown() method, and maintain a reference to that object, or let the 
init() method return a tuple of the mina object and the policy server. None of that was all that
appetising, and this is all done only on developer machines anyway, so restarting the policy server
when another node connects shouldn't be the end of the world.

So, no configuration changes required, if you need to test peer stuff locally, it should all work
now with the r124 flash player.