A first example of the third part of Le Grande GWT Refactor wherein GWT services and data classes are moved into .gwt packages.
Currently, everything is in web/client and web/data, so any page that wants
anything from that package has to know about *everything*. Once I'm done with
this giant rearrangement, then a page that is just displaying forum messages,
for example, will only need to "see" forum/gwt and a few global classes. It
won't have to know about groups, persons, items, games, rooms, ratings,
trophies or dozens of other totally unrelated things that it currently does.

GWT already does dead code removal, so we don't *really* need to do this. But
the point is that GWT has to read through hundreds of thousands of lines of
code to figure out that it doesn't need 99% of it. If we can only require that
the compiler read through what is actually relevant to the problem at hand,
things will compile a lot faster. And it will also keep us honest about what
sort of strings of spaghetti we might otherwise unthinkingly string through the
code when someone asks for a feature that results in a coupling between two
totally unrelated areas of code.