Oops, I spoke too soon. *Now* we've got all the strings of spaghetti properly placed into the right bowls and it is possible to compile a bare bones GWT app that includes only the following (ostensibly "global") code:
  <source path="org/gwtwidgets/client"/>
  <source path="com/threerings/io"/>
  <source path="com/threerings/util"/>
  <source path="com/threerings/presents/dobj"/>
  <source path="com/threerings/gwt/ui"/>
  <source path="com/threerings/gwt/util"/>

  <source path="com/threerings/msoy/data/all"/>
  <source path="com/threerings/msoy/web/data"/>
  <source path="com/threerings/msoy/web/client"/>

  <source path="client/images"/>
  <source path="client/shell"/>
  <source path="client/ui"/>
  <source path="client/util"/>

and it takes a whopping 10 seconds to compile that bare bones app which ends up
being 776 bytes in size because it does precisely nothing. So each individual
page will take at least 10 seconds to compile, probably more, but that's not
the end of the world. That's only the same amount of time it takes to compile
all 785 files in the MSOY Java code base.

Yay!