Bah, it'll be confusing that the playerIds aren't in the right order, so do the much more verbose unboxing and listifying and use that.
I also noticed that playerOidsToMemberIds() was filtering non-guests in a very
assumptive manner, so I made that use the proper guest detection code and be
explicit about what it's doing.

However, all of this should change because it relies on the players still being
online when the game is ended which definitely may not be the case. One more
reason why we should deal everywhere in memberId rather than playerOid.