We can now navigate around our inner frames and things are properly passed up to the top-level frame where we update the browser location and then pass things back down to the iframe.
Ray will be pleased to know that I managed to accomplish this while preserving
the use of hrefs so that you can still middle click to open any hyperlink in a
new window and the whole framed business will be properly reconstructed and the
URLs even look sensible while you're hovering over them. This required some
serious over-backwards-bending.