Setting stuff up for the Passport GWT page.  Badge and its extensions moved to a new com.threerings.msoy.badge.data.all package, and get included in the me and msoy GWT builds.  Also, Badge no longer depends on BadgeType, which was necessary for GWT integration.
This refactoring has completely sold me on Eclipse + Subclipse.  In the actionscript side of things
(which was thankfully the simpler refactor), I had to move the files by hand, and then go and 
compile and find all the now broken imports and change them by hand.  

In Java (Eclipse), I simply told it I wanted to move these three classes to a new package and 
*voila*, I have a new package with the classes moved, and all imports fixed for me, including the 
necessary SVN shuffling.  *Nice*