Our logon and authentication process is a labyrinthine maze of twisty passages.  The number of different ways that someone can get to the point where they have a game lobby or game instance in front of their noses is a little bit insane.  I think I thought I knew what was wrong a good 4 or 5 different times before I actually had it all figured out, and really knew what was wrong.
This takes care of the null guest name issue in games.  The meat of the matter is that now
MsoyAuthenticator generates its guest names directly off the of the guest memberId (which is 
safe across peers - although that wasn't the issue here), and it exposes the method in a 
public static method.  MsoyGameAuthenticator checks if the creds passed in contain a valid Name
(it used to just do String.valueOf(creds.getUsername()) - this being the source of the null), and 
if it doesn't, generates a guest name from MsoyAuthenticator's method with the guest's id (which 
the game server is getting just fine.  This all should lead to consistent guest names across the UI
in any place that a guest name might appear.

There was talk of potentially kicking a release out late tonight if this gets fixed so that over
the weekend the Corpse Craft players have something more interesting to look at in their multiplayer
games than a bunch of "null".  If the powers that be still think this is a good idea, I'll stay up
and kick one off around 2am or later.  If I don't hear anything before 12:30 or 1, I'll probably
stop checking and forget about releasing this for now.