How many new serialization schemes can I write in one day? We need to send WebCreds and SessionData between frames so that a page can log in and then send the resulting session information up to the containing frame (trying to do logon back and forth between the frame and page turns out to be a fiasco because we need to handle logon failure). Actually we already had this third serialization scheme already for WebCreds because when a page starts up it has to get the current creds from the frame. I just tidied it up and made it work for SessionData as well.
Just in case those of you following along at home think I'm doing this just for
fun, we can't use GWT's serialization for reasons I've already outlined in a
previous email and we can't use the serialization that we use between Flash and
GWT because that relies on JavaScriptObject and native JavaScript marshalling
and unmarshalling code but these classes are shared between the Java server and
the GWT client and thus cannot have native JavaScript methods nor make use of
JavaScriptObject. Yay.