Communicate our assigned guest id using a FlashEvent so that it gets communicated to all the appropriate parties.
With this, the split client should now be functionally complete. Yay!

There's a small annoyance with the frame flashing before displaying a world
page (ie. no content), but I'm going to switch fully over to the split client
and start the second phase of cleanup before I go back and fix that because it
requires a smidgen more refactorama.