Holy Underwear, our chat system is complicated. And fragile. There's a lot of things that do the right thing as a side effect, or "just know" certain things, like oh we're having the place name set but if *magic thing* is true, then we change it to "Game" instead. I actually changed that particular one a little bit here. Anyway, sorted out some stuff so that in gamestub mode we can use the default overlay to display game chat. It works nicely, as the tabs are used just like in a room, and other chat channels just work.
Originally, I think we planned to make it so that you could on-the-fly
switch between this and a popup. Now: I don't want to do the popup.
Fiasco! I dunno, maybe I could make it work.
But it seems like if you're pressed for room there won't be a place
to put the pop-up. What we might need though, is a button to clear
chat.