Trying to nail down this new black window style. - Nathan got it working, but he put special hacky children inside the panel to do some of the styling. - Bruno's been trying to get it working too. - I need this anyway, so I figured I'd sort it out once and for all.
Well, I have, and the current design as specified by the artists
is infeasible. I was going to argue with them anyway about
the titlebar, which I do not think should be transparent and funky.
It should look like a normal titlebar.
That would allow me to do this business with the borders and get
everythng working, which I can't do now because specifying a border
color makes the title area fill with *that* color behind the
backgroundImage.