More functional programming. Fixed long standing annoyance of mine which is that we ask for the ItemDetail and then immediately turn around and ask for the item's tags. Now we we just include the tags in the ItemDetail.
Small note for the attentive. We would like to do this:

  return Lists.transform(_groupRepo.getTagRepository().getTags(groupId), TagNameRecord.TO_TAG);

but we can't, because Google Collections is smart and returns a list that
lazily transforms its underlying elements only the first time they are
accessed. But that magical list cannot be sent to GWT because it knows not of
the wily machinations therein (and indeed it would be neither smart nor
possible to send the TagNameRecords over the wire and transform them on the
client). Instead we have to do this:

  List<TagNameRecord> trecs = _groupRepo.getTagRepository().getTags(groupId);
  return Lists.newArrayList(Iterables.transform(trecs, TagNameRecord.TO_TAG));

Which creates an ArrayList, and ArrayLists are cooties GWT can handle.