Well, I'm convinced.  Google collections is basically magical.  I got to think up what I wanted in terms of sets and unions, and then write the code just like that.
What the client really needs is:
 * A list of InProgressBadges that it can cycle through for the "Go for these new stamps" section.
 * A map of StampCategory to the Badges within it that the player has unlocked.  In the case where
   the player has both an EarnedBadge and an InProgressBadge for a given Badge series, the client
   needs the InProgressBadge so that it can display the progress for the next level.

I'll know soon enough, when the client bits get fleshed out, but this should do the above.  My 
only remaining concern is that this will feed up the badges unordered.  However, we've never 
discussed an ordering for them, so I'm inclined to leave it that way for now.  Maybe we'll sort
them in each category by progress, or by name or something, but that can be done on the client
and decided later.