- My logic chip apparently needs an upgrade, because this was totally broken.   All party games are multiplayer. - Removed 'final' that kyle slapped in there, just because it added to the   line length.
I don't know. Final, in java, is overloaded. It is sort of like "const"
and that's cool, but declaring a variable final won't stop anyone from
fudging up the internal workings of that object. Also, final has the side
effect of importing these variables into any anonymous classes created in
scope. Maybe if there were a compile-time annotation, @Const, that prevented
reassigning to a variable, I'd be all for it. But final does two things, so
I'm not the hugest fan.