A few things:
1. Don't call super() manually. There's no mystery about whether or not the
default superclass constructor is going to get called.

2. Try to use hierarchical styles: style the container and then style the
individual elements as necessary. It allows for more flexibility when doing
things in your stylesheet. In this case, we can style the container as
"favorite" and then add a selector in our CSS to say that any toggle button
inside our .favorite needs special styling. We can also then avoid adding
custom styles to any of our elements.

3. Don't put styling in the code, use the stylesheets. Setting the width to
60px turned out to be unnecessary, but should not be done in code if ever it is
necessary. It should be done in the ".favorite" CSS block. Setting horizontal
alignment should also be done in the stylesheet (there are times when one can't
avoid using code, but this is not one of those times). Setting the vertical
alignment also turned out not to be necessary.

4. Don't put your style class names in constants. They are usually only ever
used once in the code and it's indirect and verbose to have them down at the
bottom of the classfile in a constant.

5. Don't use HorizontalPanel or VerticalPanel if you can avoid it. Those wrap
everything in tables and tables bring along a whole host of annoyances and
problems that are best avoided unless there's no other way to accomplish what
you need.

6. There are times when a named inner class is useful to clarify the code, but
this was not one of those times. If you do find yourself in a situation where
the code in a ClickListener is large and unweildy, then keep the click listener
declaration inline and just call a method. For example:

    toggle.addClickListener(new ClickListener() {
        public void onClick (Widget sender) {
            setFavorite(((ToggleButton) sender).isDown());
        }
    });

    protected void setFavorite (final boolean favorite)
    {
        _itemsvc.setFavorite(_item.getIdent(), favorite, new MsoyCallback<Void>() {
            public void onSuccess (Void result) {
                _memberItemInfo.favorite = favorite;
            }
        });
    }                                  

This keeps incidental information like the handling of a button click near the
button declaration and cleanly extracts the (potentially more complex) action
being taken as a result of that click into a method that knows and says nothing
about buttons or clicking.