Reason #6,735,212 why ActionScript is fucked up:
It is not block scoped, like you might expect from a civilized language. It is
"function scoped" (and probably not even that given other special cases). So
you might write the following code thinking it was legal:

function myFunction () :void {
    for each (var tickVal :Number in ticks) {
        const diff :Number = Math.abs(tickVal - value);
        // ...
    }
}

However, it's not legal. As far as ActionScript is concerned what you actually
wrote was:

function myFunction () :void {
    const diff :Number;
    for each (var tickVal :Number in ticks) {
        diff = Math.abs(tickVal - value);
        // ...
    }
}

which is clearly not legal. They made the awesome choice to let you declare
variables anywhere you like even though you can now easily confuse yourself as
a result of their fucked up scoping.

The silver lining is that at least the compiler catches this error now instead
of just blindly letting you reassing to a const value. The mind boggles.