Step 2: Create a system for deferring startup notifications until the client can handle them.  The initial plan was to re-implement a system for storing some (or all) notifications in a set on the member object, and having the director acknowledge notifications as it displays them.
Ray and I hashed it out on awhile and settled on this.  It's annoying to make Notifications into
dset entries because they don't have a natural key.  It is also unsure what should be done
when a member moves servers with regard to stored notifications.

So, any notifications that server code knows the client may miss due to the fact that it has 
possibly not finished getting ready yet should be dispatched on the member object as normal, and 
also stored in the MemberObject.deferredNotifications list.  When the client's NotificationDirector
is ready and listening, it makes a service call to have the server dispatch any notifications 
currently sitting in that queue, when it will definitely hear them.