*sigh*.  This would have been done nearly an hour ago if not for those last fiddly 2px.  The bottom "See all" section is a few pixels taller than it should be, but if I make it any smaller, some things break.
I need to get a new icon image from Jon, but that should be a quick fix for him, which means that
with this commit, I officially declare Passport ready for general consumption!

PS: this is the revamp of the "what's next?" section of the Me page so that it contains 4 randomly
selected in progress badges, with a link to your passport page to view the rest.