A number of things:
- there is one true Log class for all of MSOY, thou shalt not have others
before it
- don't use Log.log.foo(), that's not how it's done; importing toybox.Log is
even more not how it's done; statically import msoy.Log.log and write
log.foo(), etc. in the actual code
- dispatching a notification directly on the member object in badgeAwarded
instead of going through the notification manager is just the sort of end-run
around proper encapsulation that causes things to mysteriously not work when
someone, say, modifies the notification manager to queue up pending
notifications and dispatch them all when the member has reported that they are
ready to receive them
- speaking of proper encapsulation, having the badge manager reach in and add
pending notifications directly to the queue and burdening the caller of
awardBadges() with knowing whether or not this was the right thing to do was
the wrong thing to do; the notification manager knows when to queue and when to
dispatch directly, so let it make use of that knowledge and always go through
the notification manager to dispatch notifications
- actually made the notification manager aware of when it should queue
notifications and when it should dispatch them directly
- fixed a bug with the badge manager which would result in it awarding a coin
payout for badges that had already been earned if for some reason it tried to
award a badge that already existed in a member object's set at the target level

I debated not having deferredNotifications in MemberObject at all and
maintaining an internal mapping in NotificationManager for deferred
notifications, but the free lifecycle management that comes along with
MemberObject was too hard to pass up; maybe we should have some sort of
member-local storage service that is like thread-local storage so that server
code that needs to do things like this can do so easily and without having to
worry about cleaning up after members that never show up, etc.; another one for
the list.