The maze of passages, they are twisty.
In an embed, we cannot rely on having GWT assign us a guest id and tell us
which server host and port to which to connect. So instead we have a special
servlet that the Flash client loads to find out the hostname and port of the
server to which it should connect and which also assigns it a guest id.

This was first done for games, then someone went in and added it for scenes as
well. But they didn't coordinate those activities. So when an embed servlet was
made for games, the client started up, made two separate /embed servlet
requests. The game related request assigned it a guest id and the world related
request did not (not that it would have mattered if it did). The world client
then proceeded to log on with no guestId (which is legal, the server will just
assign you a guestId when you logon), but that guestId differed from the
guestId used on the game server, so we were not able to coordinate the activity
of those two connections and the game server connection did not show up
anywhere in Whirled as someone playing a game (nor would the player have
received any earned flow if they registered, etc. etc.).

Now if we're starting up and going into a game lobby, we let the game liaison
make the /embed request, get the host and port of the game server it needs to
connect to *and* the server that is hosting that game's whirled, and a guest
id. It then tells the world server what server it will be connecting to and the
guestId to use when doing so, so that it can get into the game's whirled
maximally efficiently and using the right guest id so that everything is
coordinated.

If we're in our GWT client, everything is much simpler because GWT can just
tell Flash beforehand what its guestId is and which servers to go to and both
the world and game clients can connect directly to where they need to go with
the right guestId. However, we're now taking advantage of the information we
get back with the game LaunchConfig in GWT and avoiding a second round trip to
the server to obtain a default server host/port for the Flash client. We just
use the game's whirled's hosting server's host and port (boy that's a lot of
overloaded terminology) because we'll be going to that server once the lobby is
loaded anyway.